Update main.cpp
Added a more complete ROS thread. The thread was tested separately from the rest of the program by alec and works by itself but in the main program the ROS library causes problems. I believe alec compiled the ROS program and library with the old mbed.h and not the mbed-os library which may explain where the issues stem from. 

Also added code to parse out different pgns so that ROS can push them to the PI. Not yet successful at converting the pgn data back into a useable float